# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BizOpt API
  description: |
    REST API for BizOpt web and native (iOS) clients.
    Authenticate with `Authorization: Bearer <accessToken>` from `/api/auth/mobile/login`.
    Device-bound routes also require `x-device-fingerprint`, `x-device-label`, `x-device-platform`.
  version: 1.0.0
  contact:
    name: BizOpt
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://{host}
    description: Production
    variables:
      host:
        default: app.example.com
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    MobileLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        deviceLabel:
          type: string
    MobileTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        tokenType:
          type: string
          example: Bearer
        userId:
          type: string
        orgId:
          type: string
          nullable: true
...
paths:
  /api/auth/mobile/login:
    post:
      tags: [Auth]
      summary: Mobile login (email + password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileLoginRequest'
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...